requirements.txt
snowflake-connector-python>=3.6.0
pyyaml>=6.0
streamlit>=1.30.0
altair>=5.0.0
pandas>=2.0.0
